BigData_project/requirements.txt
cassandra-driver
pandas
prettytable
